MTN: rename search to fuzzybroker and make a module for auto-config fuzzydatabroker
